requirements.txt
streamlit
streamlit-options-menu
Pillow